requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
joblib>=1.1.0

# API and Web
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
requests>=2.28.0

# Dashboard
streamlit>=1.20.0
plotly>=5.13.0

# Database
psycopg2-binary>=2.9.0
neo4j>=5.5.0

# Data Processing
PyYAML>=6.0
python-dateutil>=2.8.0
tqdm>=4.64.0

# Scheduling
schedule>=1.1.0

# Medical/Health specific
lifelines>=0.27.0

# Utilities
python-multipart>=0.0.5

#quick start

python run_demo.py
python api/api_server.py          # API at localhost:8000
streamlit run dashboard/health_dashboard.py  # Dashboard at localhost:8501
